\name{sin1reg}
\alias{sin1reg}
\docType{data}
\title{sin1 regression dataset}
\description{
Simple synthetic dataset with 1000 points, where y=0.7*sin(pi*x1/2000)+0.3*sin(pi*x2/2000)
}
\usage{data(sin1reg)}
\format{
  The format is:
 chr "sin1reg"
}
\details{Simple synthetic dataset with 1000 points, where y=0.7*sin(pi*x1/2000)+0.3*sin(pi*x2/2000)}
\source{To be filled... }
\references{To be filled... }
\examples{
data(sin1reg)
print(summary(sin1reg))
}
\keyword{datasets}
