\name{rminer-internal}
\alias{AAD_responses}
\alias{aggregate_imp}
\alias{agg_matrix_imp}
\alias{balanced_responses}
\alias{cmatrixplot}
\alias{datalevels}
\alias{INTERPOLATE}
\alias{LIFTcurve}
\alias{MCrandom}
\alias{mids}
\alias{RECcurve}
\alias{ROCcurve}
\alias{ranges}
\alias{rmboxplot}
\alias{rmsample}
\alias{rmtable}
\alias{s_measure}
\alias{TPR_FOR_FPR}
\alias{addfactor}
\alias{attcorrelated}
\alias{avg_imp}
\alias{avg_imp_1D}
\alias{bestfit}
\alias{bssearch}
\alias{conflevel}
\alias{Conf}
\alias{crossfolds}
\alias{curvearea}
\alias{defaultask}
\alias{defaultfeature}
\alias{defaultmodel}
\alias{defaultmpar}
\alias{defaultsearch}
\alias{dforder}
\alias{dlplot}
\alias{enlarge}
\alias{factorize}
\alias{factor2numeric}
\alias{feature_needed}
\alias{fenlarge}
\alias{filter_equal}
\alias{forplot}
\alias{getmetric}
\alias{gradient_responses}
\alias{hotdeck}
\alias{impvalue}
\alias{invtransform}
\alias{isbest}
\alias{knn.fit}
\alias{majorClass}
\alias{mean_resp}
\alias{meanint}
\alias{medianfirst}
\alias{middleclass}
\alias{missingatts}
\alias{mlp.fit}
\alias{modelnamespar}
\alias{modelnpar}
\alias{modelplot}
\alias{mostcommon}
\alias{mpause}
\alias{naivebayes}
\alias{naivebayes.default}
\alias{naivebayes.formula}
\alias{one_of_c}
\alias{output_index}
\alias{partialcurve}
\alias{pathlength_responses}
\alias{plotH}
\alias{predict.naivebayes}
\alias{print.naivebayes}
\alias{range_responses}
\alias{resp_to_list}
\alias{rminerhistogram}
\alias{scaleinputs}
\alias{scaleinputs2}
\alias{svm.fit}
\alias{transform_needed}
\alias{trap_area}
\alias{tsacf}
\alias{tsplot}
\alias{twoclassLift}
\alias{twoclassROC}
\alias{uniform_design}
\alias{variance_responses}
\alias{vaveraging}
\alias{worst}
\alias{xmiddle_point}
\alias{xtransform}
\alias{yaggregate}
\title{Internal rminer Functions}
\description{Internal rminer functions
}
\details{
These are not to be called by the user (or in some cases are just
waiting for proper documentation to be written :).
}
\keyword{internal}
