\name{rminer-internal}
\alias{Accuracy}
\alias{Aliftperc}
\alias{Auc}
\alias{Brier}
\alias{Correlation}
\alias{Gmad}
\alias{Gmean}
\alias{Gmrae}
\alias{Gmse}
\alias{INTERPOLATE}
\alias{Kappa}
\alias{LIFTcurve}
\alias{MMinkowski}
\alias{Mad}
\alias{Mape}
\alias{Mase}
\alias{MdMinkowski}
\alias{Mdae}
\alias{Mdape}
\alias{Mdrae}
\alias{Mdse}
\alias{Me}
\alias{Mrae}
\alias{Mse}
\alias{Nalift}
\alias{Narec}
\alias{Nauc}
\alias{R2}
\alias{R22}
\alias{RECcurve}
\alias{ROCcurve}
\alias{Rae}
\alias{Rmdspe}
\alias{Rmse}
\alias{Rmspe}
\alias{Rrse}
\alias{Rse}
\alias{SMinkowski}
\alias{Sad}
\alias{Smape}
\alias{Smdape}
\alias{Ss}
\alias{Sse}
\alias{TPR_FOR_FPR}
\alias{Tbrier}
\alias{TheilsU2}
\alias{Tolerance}
\alias{Tprfpr}
\alias{addfactor}
\alias{attcorrelated}
\alias{avg_imp}
\alias{avg_imp_1D}
\alias{bestfit}
\alias{bssearch}
\alias{classification.metrics}
\alias{conflevel}
\alias{Conf}
\alias{crossfolds}
\alias{curvearea}
\alias{defaultask}
\alias{defaultfeature}
\alias{defaultmodel}
\alias{defaultmpar}
\alias{defaultsearch}
\alias{dforder}
\alias{dlplot}
\alias{enlarge}
\alias{Error}
\alias{factor2numeric}
\alias{feature_needed}
\alias{fenlarge}
\alias{filter_equal}
\alias{forplot}
\alias{getmetric}
\alias{gradient_responses}
\alias{guess_k}
\alias{hotdeck}
\alias{impvalue}
\alias{invtransform}
\alias{isbest}
\alias{knn.fit}
\alias{majorClass}
\alias{mean_resp}
\alias{meanint}
\alias{medianfirst}
\alias{middleclass}
\alias{missingatts}
\alias{mlp.fit}
\alias{modelnamespar}
\alias{modelnpar}
\alias{modelplot}
\alias{mostcommon}
\alias{mpairwise}
\alias{multiROC}
\alias{mypause}
\alias{naivebayes}
\alias{naivebayes.default}
\alias{naivebayes.formula}
\alias{nominal}
\alias{one_of_c}
\alias{output_index}
\alias{partialcurve}
\alias{pathlength_responses}
\alias{plotH}
\alias{predict.naivebayes}
\alias{print.naivebayes}
\alias{randomwalk}
\alias{range_responses}
\alias{readts}
\alias{regression.metrics}
\alias{resp_to_list}
\alias{rminerhistogram}
\alias{scaleinputs}
\alias{scaleinputs2}
\alias{substituteNV}
\alias{svm.fit}
\alias{tolerance}
\alias{transform_needed}
\alias{trap_area}
\alias{tsacf}
\alias{tsplot}
\alias{twoclassLift}
\alias{twoclassROC}
\alias{uniform_design}
\alias{variance_responses}
\alias{vaveraging}
\alias{worst}
\alias{xmiddle_point}
\alias{xtransform}
\title{Internal rminer Functions}
\description{Internal rminer functions
}
\details{
These are not to be called by the user (or in some cases are just
waiting for proper documentation to be written :).
}
\keyword{internal}
